ai-server/requirements.txt
# Meat-A-Eye AI Server Requirements
# 기존 환경에 이미 설치된 패키지 기준으로 작성

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
efficientnet_pytorch>=0.7.1

# Web Server
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.1

# OCR
easyocr>=1.7.0

# Utilities
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
